FastReg a function to perform fast simple linear or logistic regression

FastReg a function to perform fast simple linear or logistic regression

FastReg(
  phenotype = "bin.resp",
  regression.type = "logistic",
  Pvalue.dist = "t.dist",
  output.exclude.covar = TRUE,
  maf.threshold = 1e-13,
  hwe.threshold = 1e-13,
  no.intercept = FALSE,
  colinearity.rsq = 1,
  poi.block.size = 0,
  max.iter = 6,
  rel.conv.tolerance = 0.01,
  abs.conv.tolerance = 0.01,
  max.openmp.threads = 1,
  pheno.file = "testdata_1k_by_5k.bin.pheno.txt",
  pheno.rowname.cols = "ind",
  pheno.file.delim = "tab",
  covar.file = "testdata_1k_by_5k.covar.txt",
  covar.rowname.cols = "ind",
  covar.file.delim = "tab",
  POI.file.dir = "testdata/",
  POI.file.delim = "tab",
  POI.file.format = "h5",
  POI.type = "dosage",
  POI.effect.type = "additive",
  covariates = c("age", "sex", "eth", "treatment", "severity"),
  covariate.type = c("numeric", "categorical", "categorical", "categorical",
    "categorical"),
  covariate.standardize = c(TRUE, FALSE, FALSE, FALSE, FALSE),
  covariate.levels = c("", "F,M", "afr,asi,eur", "Placebo,Test",
    "Very Low,Low,Moderate,High,Very High,Extreme"),
  covariate.ref.level = c("", "F", "eur", "Placebo", "Very Low"),
  POI.covar.interactions = c(""),
  split.by = c(""),
  output.dir = "test",
  compress.results = FALSE,
  max.workers = 0
)

Arguments

phenotype

bin.resp | num.resp. Default: bin.resp

regression.type

logistic | linear. Default: logistic

Pvalue.dist

t.dist | norm.dist. Default: t.dist

output.exclude.covar

TRUE | FALSE. Default: TRUE. Excludes covariates in the results output

maf.threshold

Default: 1e-13. Minor Allele frequency threshold. Acceptable values: 0-0.5

hwe.threshold

Default: 1e-13. Hardy-Weinberg Equilibrium threshold. Acceptable values: 0-0.5

no.intercept

TRUE | FALSE. Default: FALSE.

colinearity.rsq

Default: 1.0. Filter covariates if above R-squared colinearity threshold. Acceptage values: 0.8-1.0

poi.block.size

Default: 0. Overrides auto-calculated poi block size.

max.iter

Default: 6. Number of logistic regression iterations. Doesn't apply when regression.type = linear.

rel.conv.tolerance

Default: 0.01. Relative convergence threshold for POIs.

abs.conv.tolerance

Default: 0.01. Absolute convergence threshold for POIs.

max.openmp.threads

Default: 1. Overrides number of threads used. By default FastReg will use 1 openmp threads.

pheno.file

relative path to phenotype file. Contains outcomes that will be modelled.

pheno.rowname.cols

Default: ind. column to be treated as the subject identifier when matching across files. Can be multiple columns (comma separated).

pheno.file.delim

tab|space|comma. Default: tab. delimiter used in phenotype file.

covar.file

relative path to covariates file.

covar.rowname.cols

Default: ind. column to be treated as the subject identifier when matching across files. Can be multiple columns (comma separated).

covar.file.delim

tab|space|comma. Default: tab. delimiter used in covariate file.

POI.file.dir

relative path to the directory with POI file(s).

POI.file.delim

tab|space|comma. Default: tab. delimiter used in POI file if POI.file.format != h5

POI.file.format

h5|txt. Default: h5. POI data file format.

POI.type

genotype|dosage. Default: dosage.

POI.effect.type

additive|dominant|recessive. Default: additive.

covariates

Default: c("age","sex","eth","treatment","severity"). List of column names of the covariates file.

covariate.type

Default: c("numeric","categorical","categorical","categorical","categorical"). List of column data type for each covariate listed in covariates.

covariate.standardize

Default: c(TRUE,FALSE,FALSE,FALSE,FALSE). List of booleans representing which covariates are to be standardized.

covariate.levels

Default: c("", "F,M","afr,asi,eur","Placebo,Test","Very Low,Low,Moderate,High,Very High,Extreme"). List of unique values for categorical columns. (comma separated for each unique value)

covariate.ref.level

Default: c("","F","eur","Placebo","Very Low"). List of reference level for covariate value (comma separated)

POI.covar.interactions

Default: c(""). List of poi-covariate interactions

split.by

Default: c(""). List of covariates values to split by when stratifying the data.

output.dir

Default: results. Relative path to directory for output. Directory will be created if it doesn't exist

compress.results

TRUE|FALSE. Default: FALSE. Compress results inside the output directory.

max.workers

Default: 0. Number of processes FastReg will spawn. By default FastReg will spawn min(# of poi files, # of available workers)

Value

numeric denoting elapsed.time