create_test_dataset function to generate test dataset for evaluation of FastReg

create_test_dataset function to generate test dataset for evaluation of FastReg

create_test_dataset(
  num.poi = 50000,
  num.ind = 5000,
  seed = 12133,
  coeff.sd = 0,
  bin.resp.mean = 0.2,
  num.resp.mean = 24,
  num.resp.sd = 5,
  poi.type = "genotype",
  poi.chunk.size = 100,
  poi.compression.level = 7,
  data.dir = ".",
  data.num.chunks = 10,
  prefix = "testdata_5k_by_50k",
  covariates = c("age", "sex", "treatment"),
  poi.file.type = "txt",
  verbose = TRUE
)

Arguments

num.poi

an integer denoting number of predictors of interest (SNP calls)

num.ind

an integer denoting number of individuals

seed

an integer denoting random number generate seed

coeff.sd

numeric value denoting the standard deviation parameter for regression coefficients for covariates (default=0, regression coefficients for covarites are set to zero)

bin.resp.mean

numeric value between 0 and 1 denoting baseline incidence rate for binary response

num.resp.mean

numeric value denoting overall mean of numeric response

num.resp.sd

numeric value denoting overall standard deviation of numeric response

poi.type

character must be either "genotype" or "dosage"

poi.chunk.size

an integer denoting poi chunk size used during H5 file generation

poi.compression.level

an integer denoting compression level used during H5 file creation

data.dir

output directory

data.num.chunks

number of files to split the data into

prefix

prefix for all files created

covariates

list of covariates to generate. Supported types are: age, sex, treatment, eth, severity

poi.file.type

'h5' or 'txt'

verbose

logical to control display of progress messages (default=TRUE)

Value

a list consisting of dataset size (num.poi, num.ind) as well as regression coefficients used for both binary and numeric response